Adds view functionality
Gives Jenkins Job Builder the ability to work with views using new methods in
python-jenkins to work with views.
Views are able to be created, updated, and deleted.
New modules for List view and Build Pipeline view are added.
Entry points are created for both respectively in a new section for views.
Example Yaml and Xml files are also included for each module in folders for
testing both build pipeline and list views.
There are also new tests for testing the deletion of views.

+++ jenkins_jobs/cmd.py
@@ -124,11 +124,26 @@ def create_parser():
                                help='colon-separated list of paths to'
                                     ' YAML files or directories')
 
+    # subparser: delete-view
+    parser_delete_view = subparser.add_parser('delete-view',
+                                              parents=[recursive_parser])
+    parser_delete_view.add_argument('name', help='name of view', nargs='+')
+
     # subparser: delete-all
-    subparser.add_parser('delete-all',
-                         help='delete *ALL* jobs from Jenkins server, '
-                         'including those not managed by Jenkins Job '
-                         'Builder.')
+    parser_delete_all = subparser.add_parser('delete-all',
+                                             help='delete *ALL* jobs from '
+                                             'Jenkins server, including those '
+                                             'not managed by Jenkins Job '
+                                             'Builder.')
+    parser_delete_all.add_argument('-j', '--only-jobs',
+                                   help='delete only Jobs',
+                                   action='store_true', dest='only_jobs',
+                                   default=False,)
+    parser_delete_all.add_argument('-v', '--only-views',
+                                   help='delete only Views',
+                                   action='store_true', dest='only_views',
+                                   default=False,)
+
     parser.add_argument('--conf', dest='conf', help='configuration file')
     parser.add_argument('-l', '--log_level', dest='log_level', default='info',
                         help="log level (default: %(default)s)")
@@ -283,11 +298,30 @@ def execute(options, config):
     if options.command == 'delete':
         for job in options.name:
             builder.delete_job(job, options.path)
+    elif options.command == 'delete-view':
+        for view in options.name:
+            builder.delete_view(view, options.path)
     elif options.command == 'delete-all':
-        confirm('Sure you want to delete *ALL* jobs from Jenkins server?\n'
-                '(including those not managed by Jenkins Job Builder)')
-        logger.info("Deleting all jobs")
-        builder.delete_all_jobs()
+        deleting_jobs = False
+        deleting_views = False
+        if options.only_jobs == options.only_views:
+            reach = 'jobs AND views'
+            deleting_jobs = True
+            deleting_views = True
+        elif options.only_jobs and not options.only_views:
+            reach = 'jobs'
+            deleting_jobs = True
+        elif options.only_views and not options.only_jobs:
+            reach = 'views'
+            deleting_views = True
+        confirm('Sure you want to delete *ALL* %s from Jenkins server?\n'
+                '(including those not managed by Jenkins Job Builder)', reach)
+        if deleting_jobs:
+            logger.info("Deleting all jobs")
+            builder.delete_all_jobs()
+        if deleting_views:
+            logger.info("Deleting all views")
+            builder.delete_all_views()
     elif options.command == 'update':
         logger.info("Updating jobs in {0} ({1})".format(
             options.path, options.names))
